feat: criar utils para mapicon
